Started to implment themes and modifyied the renamed the NumericUpDown user control
Themes are partially implmented, however you cant dynamically switch between the themes and they are not fully implmented.

I also renamed the user control "NumericUpDown" to be "NumericEntryControl" and modified the binding of the Textbox so it doesn't need to have a data context for the user control, as shown in where I found where to make this usercontrol.