Tool tips implmented and mis alterations
Now there are tooltips for the scene name and the custom frames input fields.
Also modified the Validators class so it will return a instance of the class "ValidatorsReturn" which will also include an error message for what cause the error.
Next, I replaced the "Loaded" event listener with "Initialized" for the main window so that it will not have the main window popup breafly when opening the browse window and it will have the theme applied to said window.
Lastly, all the other windows will now have the background theme applied to them.  HOWEVER, they do not share the theme of the main window. Example the text color stays black regardless on if the theme is in light or dark mode.