Errors and tooltips colors are now defined in the theme files. Read description for additional changes.
Next, I fixed an issue with the labels for both the scene name and custom frames input not being lined up with the box.
Next, I made a tiny change to the render section for the input labels so they are more seperated than they were before.
Lastly, I changed the error tree messages so they better describe the issues found.