Fixed a issue with rendering output folder and removed "Frame Range" option from the rendering type
This set of changes also formats the outputed file so it is easier to read if you open it for editing.

It also should fix the issue with selecting the shutdown option, but this will most likely going to be changed for a new command as talked about in the code.